revert most table changes
