+++ packages/react-table/src/components/Table/utils/decorators/cellWidth.ts
@@ -1,6 +1,6 @@
 import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/react-styles/css/components/Table/table';
-import { ITransform } from '../../base';
+import { ITransform } from '../../Table';
 import { capitalize } from '../utils';
 
 type WidthClass =
