+++ packages/react-table/src/components/Table/utils/decorators/collapsible.tsx
@@ -3,7 +3,7 @@ import { css } from '@patternfly/react-styles';
 import styles from '@patternfly/react-styles/css/components/Table/table';
 import { CollapseColumn } from '../../CollapseColumn';
 import { ExpandableRowContent } from '../../ExpandableRowContent';
-import { IExtra, IFormatterValueType, IFormatter, decoratorReturnType } from '../../base';
+import { IExtra, IFormatterValueType, IFormatter, decoratorReturnType } from '../../Table';
 
 export const collapsible: IFormatter = (
   value: IFormatterValueType,
