+++ packages/react-table/src/components/Table/utils/transformers.tsx
@@ -10,7 +10,8 @@ export { compoundExpand } from './decorators/compoundExpand';
 export { headerCol } from './decorators/headerCol';
 export { classNames, Visibility } from './decorators/classNames';
 export { info } from './decorators/info';
-import { IFormatterValueType, IExtra, ITransform } from '../base';
+
+import { IFormatterValueType, IExtra, ITransform } from '../Table';
 
 const emptyTD: ITransform = () => ({
   scope: '',
