+++ packages/react-table/src/test-helpers/data-sets.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable no-console */
 import * as React from 'react';
-import { IRow, ICell, IActions } from '../components/Table/base';
-import { EditableTextCell } from '../components';
+import { IRow, ICell, IActions, EditableTextCell } from '../components/Table';
 
 export const columns: (ICell | string)[] = [
   { title: 'Header cell' },
